+++ src/components/common/GridDisplay.vue
@@ -16,7 +16,7 @@
 import { defineComponent, PropType } from 'vue';
 import { SpotType } from '@/types/SpotType';
 
-import '@/components/common/wc_spot-item';
+import '@/components/common/wc-spot-item';
 
 export default defineComponent({
 	name: 'GridDisplay',
