+++ src/components/common/wc-image-item.ts
