+++ src/components/search/wc-searchbar.ts
