+++ src/store/spinnerStore.ts
